#1314 Remove ParametersSetFactory as it is only a wrapper for a call to boost::shared_ptr and it creates issue when used in multi-threading in modelica models
