+++ dynawo/sources/API/PAR/PARParametersSet.cpp
@@ -13,133 +13,328 @@
 
 /**
  * @file PARParametersSet.cpp
- * @brief Dynawo parameters set : implementation for iterator
+ * @brief Dynawo parameters set : implementation file
  *
  */
+#include <sstream>
+#include <set>
+#include "DYNMacrosMessage.h"
 
 #include "PARParametersSet.h"
-#include "PARParametersSetImpl.h"
+#include "PARParameter.h"
+#include "PARParameterFactory.h"
 
+using std::map;
+using std::set;
+using std::string;
+using std::vector;
+using std::stringstream;
+
+using boost::dynamic_pointer_cast;
 using boost::shared_ptr;
+using boost::unordered_map;
 
 namespace parameters {
 
-ParametersSet::parameter_const_iterator::parameter_const_iterator(const ParametersSet::Impl* iterated, bool begin) :
-impl_(new BaseIteratorImpl(iterated, begin)) { }
-
-ParametersSet::parameter_const_iterator::parameter_const_iterator(const ParametersSet::parameter_const_iterator& original) :
-impl_(new BaseIteratorImpl(*(original.impl_))) { }
-
-ParametersSet::parameter_const_iterator::~parameter_const_iterator() {
-  delete impl_;
-  impl_ = NULL;
+ParametersSet::ParametersSet(const string& id) :
+id_(id) {
 }
 
-ParametersSet::parameter_const_iterator&
-ParametersSet::parameter_const_iterator::operator=(const ParametersSet::parameter_const_iterator& other) {
-  if (this == &other)
-    return *this;
-  delete impl_;
-  impl_ = (other.impl_ == NULL)?NULL:new BaseIteratorImpl(*(other.impl_));
-  return *this;
+const std::string&
+ParametersSet::getId() const {
+  return id_;
 }
+
+const std::string&
+ParametersSet::getFilePath() const {
+  return filepath_;
+}
+
+void
+ParametersSet::setFilePath(const std::string& filepath) {
+  filepath_ = filepath;
+}
+
+// in case of using omc, parameter A[1] will be renamed A_1
+// A[1][1] will ne renamed A_1_1
+// in par file, A[1] will be declared with row = 1, column =1, so we create two names:
+// A_1_1 and A_1 (using column index)
+
+vector<string> ParametersSet::tableParameterNames(const string& name, const string& row, const string& column) const {
+  vector<string> names;
+  stringstream nameFull;
+
+  nameFull.str("");
+  nameFull << name << "_" << row << "_" << column << "_";
+  names.push_back(nameFull.str());
+
+  // all further names will be created as aliases
+  if (row == "1") {
+    nameFull.str("");
+    nameFull << name << "_" << column << "_";
+    names.push_back(nameFull.str());
+  }
+
+  return names;
+}
+
+shared_ptr<ParametersSet> ParametersSet::createAlias(const string& aliasName, const string& origName) {
+  if ((!hasParameter(origName)) || (hasParameter(aliasName))) {
+    throw DYNError(DYN::Error::API, ParameterAliasFailed, aliasName, origName, id_);
+  }
+  parameters_[aliasName] = parameters_[origName];
+
+  return shared_from_this();
+}
+
+// only create one parameter object, then create aliases pointing towards the first object
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, bool value, const string& row, const string& column) {
+  return addParameter<bool>(name, value, row, column);
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, int value, const string& row, const string& column) {
+  return addParameter<int>(name, value, row, column);
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, double value, const string& row, const string& column) {
+  return addParameter<double>(name, value, row, column);
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, const string& value, const string& row, const string& column) {
+  return addParameter<string>(name, value, row, column);
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, const bool value) {
+  return addParameter(ParameterFactory::newParameter(name, value));
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, const int value) {
+  return addParameter(ParameterFactory::newParameter(name, value));
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, const double value) {
+  return addParameter(ParameterFactory::newParameter(name, value));
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::createParameter(const string& name, const string& value) {
+  return addParameter(ParameterFactory::newParameter(name, value));
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::addParameter(const shared_ptr<Parameter>& param) {
+  const string name = param->getName();
+  if (hasParameter(name)) {
+    throw DYNError(DYN::Error::API, ParameterAlreadyInSet, name, id_);
+  }
+  parameters_[name] = param;
+  return shared_from_this();
+}
+
+const shared_ptr<Parameter>
+ParametersSet::getParameter(const string& name) const {
+  map<string, shared_ptr<Parameter> >::const_iterator itParam = parameters_.find(name);
+  if (itParam == parameters_.end())
+    throw DYNError(DYN::Error::API, ParameterNotFoundInSet, name, id_);
+
+  itParam->second->setUsed(true);
+
+  return itParam->second;
+}
+
+const shared_ptr<Reference>
+ParametersSet::getReference(const string& name) const {
+  unordered_map<string, shared_ptr<Reference> >::const_iterator itRef = references_.find(name);
+  if (itRef == references_.end())
+    throw DYNError(DYN::Error::API, ReferenceNotFoundInSet, name, id_);
+  return itRef->second;
+}
+
+bool
+ParametersSet::hasParameter(const string& name) const {
+  return (parameters_.find(name) != parameters_.end());
+}
+
+bool
+ParametersSet::hasReference(const string& name) const {
+  return (references_.find(name) != references_.end());
+}
+
+void
+ParametersSet::extend(shared_ptr<ParametersSet> parametersSet) {
+  const map<string, shared_ptr<Parameter> >& mapParameters = dynamic_pointer_cast<ParametersSet>(parametersSet)->getParameters();
+  parameters_.insert(mapParameters.begin(), mapParameters.end());
+}
+
+vector<string>
+ParametersSet::getParametersNames() const {
+  vector<string> returnVector;
+  for (map<string, shared_ptr<Parameter> >::const_iterator itParams = parameters_.begin();
+          itParams != parameters_.end();
+          ++itParams) {
+    returnVector.push_back(itParams->first);
+  }
+  return returnVector;
+}
+
+vector<string>
+ParametersSet::getParamsUnused() const {
+  vector<string> returnVector;
+  for (map<string, shared_ptr<Parameter> >::const_iterator itParams = parameters_.begin();
+          itParams != parameters_.end();
+          ++itParams) {
+    if (!itParams->second->getUsed()) {
+      returnVector.push_back(itParams->first);
+    }
+  }
+  return returnVector;
+}
+
+vector<string>
+ParametersSet::getReferencesNames() const {
+  vector<string> returnVector;
+  set<string> orderedNames;
+  for (unordered_map<string, shared_ptr<Reference> >::const_iterator itRefs = references_.begin();
+          itRefs != references_.end();
+          ++itRefs) {
+    orderedNames.insert(itRefs->first);
+  }
+  for (set<string>::const_iterator itRefs = orderedNames.begin();
+          itRefs != orderedNames.end();
+          ++itRefs) {
+    returnVector.push_back(*itRefs);
+  }
+  return returnVector;
+}
+
+shared_ptr<ParametersSet>
+ParametersSet::addReference(boost::shared_ptr<Reference> ref) {
+  const string& refName = ref->getName();
+  if (hasReference(refName)) {
+    throw DYNError(DYN::Error::API, ReferenceAlreadySet, refName);
+  }
+  references_[refName] = ref;
+  return shared_from_this();
+}
+
+map<string, shared_ptr<Parameter> >&
+ParametersSet::getParameters() {
+  return parameters_;
+}
+
+ParametersSet::parameter_const_iterator
+ParametersSet::cbeginParameter() const {
+  return ParametersSet::parameter_const_iterator(this, true);
+}
+
+ParametersSet::parameter_const_iterator
+ParametersSet::cendParameter() const {
+  return ParametersSet::parameter_const_iterator(this, false);
+}
+
+ParametersSet::reference_const_iterator
+ParametersSet::cbeginReference() const {
+  return ParametersSet::reference_const_iterator(this, true);
+}
+
+ParametersSet::reference_const_iterator
+ParametersSet::cendReference() const {
+  return ParametersSet::reference_const_iterator(this, false);
+}
+
+ParametersSet::parameter_const_iterator::parameter_const_iterator(const ParametersSet* iterated, bool begin) :
+current_((begin ? iterated->parameters_.begin() : iterated->parameters_.end())) { }
 
 ParametersSet::parameter_const_iterator&
 ParametersSet::parameter_const_iterator::operator++() {
-  ++(*impl_);
+  ++current_;
   return *this;
 }
 
 ParametersSet::parameter_const_iterator
 ParametersSet::parameter_const_iterator::operator++(int) {
   ParametersSet::parameter_const_iterator previous = *this;
-  (*impl_)++;
+  current_++;
   return previous;
 }
 
 ParametersSet::parameter_const_iterator&
 ParametersSet::parameter_const_iterator::operator--() {
-  --(*impl_);
+  --current_;
   return *this;
 }
 
 ParametersSet::parameter_const_iterator
 ParametersSet::parameter_const_iterator::operator--(int) {
   ParametersSet::parameter_const_iterator previous = *this;
-  (*impl_)--;
+  current_--;
   return previous;
 }
 
 bool
 ParametersSet::parameter_const_iterator::operator==(const ParametersSet::parameter_const_iterator& other) const {
-  return *impl_ == *(other.impl_);
+  return current_ == other.current_;
 }
 
 bool
 ParametersSet::parameter_const_iterator::operator!=(const ParametersSet::parameter_const_iterator& other) const {
-  return *impl_ != *(other.impl_);
+  return current_ != other.current_;
 }
 
 const shared_ptr<Parameter>&
 ParametersSet::parameter_const_iterator::operator*() const {
-  return *(*impl_);
+  return current_->second;
 }
 
 const shared_ptr<Parameter>*
 ParametersSet::parameter_const_iterator::operator->() const {
-  return impl_->operator->();
+  return &(current_->second);
 }
 
-ParametersSet::reference_const_iterator::reference_const_iterator(const ParametersSet::Impl* iterated, bool begin) :
-impl_(new BaseIteratorRefImpl(iterated, begin)) { }
+// for Reference
 
-ParametersSet::reference_const_iterator::reference_const_iterator(const ParametersSet::reference_const_iterator& original) :
-impl_(new BaseIteratorRefImpl(*(original.impl_))) { }
-
-ParametersSet::reference_const_iterator::~reference_const_iterator() {
-  delete impl_;
-  impl_ = NULL;
-}
-
-ParametersSet::reference_const_iterator&
-ParametersSet::reference_const_iterator::operator=(const ParametersSet::reference_const_iterator& other) {
-  if (this == &other)
-    return *this;
-  delete impl_;
-  impl_ = (other.impl_ == NULL)?NULL:new BaseIteratorRefImpl(*(other.impl_));
-  return *this;
-}
+ParametersSet::reference_const_iterator::reference_const_iterator(const ParametersSet* iterated, bool begin) :
+current_((begin ? iterated->references_.begin() : iterated->references_.end())) { }
 
 ParametersSet::reference_const_iterator&
 ParametersSet::reference_const_iterator::operator++() {
-  ++(*impl_);
+  ++current_;
   return *this;
 }
 
 ParametersSet::reference_const_iterator
 ParametersSet::reference_const_iterator::operator++(int) {
   ParametersSet::reference_const_iterator previous = *this;
-  (*impl_)++;
+  current_++;
   return previous;
 }
 
 bool
 ParametersSet::reference_const_iterator::operator==(const ParametersSet::reference_const_iterator& other) const {
-  return *impl_ == *(other.impl_);
+  return current_ == other.current_;
 }
 
 bool
 ParametersSet::reference_const_iterator::operator!=(const ParametersSet::reference_const_iterator& other) const {
-  return *impl_ != *(other.impl_);
+  return current_ != other.current_;
 }
 
 const shared_ptr<Reference>&
 ParametersSet::reference_const_iterator::operator*() const {
-  return *(*impl_);
+  return current_->second;
 }
 
 const shared_ptr<Reference>*
 ParametersSet::reference_const_iterator::operator->() const {
-  return impl_->operator->();
+  return &(current_->second);
 }
+
 }  // namespace parameters
