+++ dynawo/sources/API/PAR/PARParametersSetCollectionImpl.cpp
@@ -20,7 +20,6 @@
 #include "DYNMacrosMessage.h"
 
 #include "PARParametersSet.h"
-#include "PARParametersSetFactory.h"
 #include "PARParametersSetCollectionImpl.h"
 
 using std::map;
