+++ dynawo/sources/API/PAR/PARXmlHandler.cpp
@@ -33,7 +33,6 @@
 #include "PARReference.h"
 #include "PARReferenceFactory.h"
 #include "PARParametersSet.h"
-#include "PARParametersSetFactory.h"
 #include "PARParametersSetCollection.h"
 #include "PARParametersSetCollectionFactory.h"
 
@@ -89,7 +88,7 @@ refHandler_(parser::ElementName(par_ns, "reference")) {
 
 void
 SetHandler::create(attributes_type const & attributes) {
-  setRead_ = ParametersSetFactory::newInstance(attributes["id"].as_string());
+  setRead_ = shared_ptr<ParametersSet>(new ParametersSet(attributes["id"].as_string()));
 }
 
 shared_ptr<ParametersSet>
