+++ dynawo/sources/Modeler/Common/DYNModelDescription.h
@@ -24,12 +24,8 @@
 #include <boost/shared_ptr.hpp>
 #include <boost/weak_ptr.hpp>
 
-#include "PARParametersSetFactory.h"
 #include "DYDModel.h"
-
-namespace parameters {
-class ParametersSet;
-}  // namespace parameters
+#include "PARParametersSet.h"
 
 
 namespace DYN {
@@ -146,7 +142,8 @@ class ModelDescription {
    * @param params : set of parameters associated to the model
    */
   inline void setParametersSet(const boost::shared_ptr<parameters::ParametersSet>& params) {
-    parameters_ = parameters::ParametersSetFactory::copyInstance(params);
+    if (params)
+      parameters_ = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet(*params));
   }
 
   /**
