+++ dynawo/sources/Modeler/Common/test/Test.cpp
@@ -40,7 +40,6 @@
 #include "DYNModelMulti.h"
 #include "DYNSubModel.h"
 #include "PARParametersSet.h"
-#include "PARParametersSetFactory.h"
 #include "DYNDynamicData.h"
 
 #include "DYNError.h"
@@ -51,7 +50,6 @@ using boost::shared_ptr;
 using boost::dynamic_pointer_cast;
 
 using parameters::ParametersSet;
-using parameters::ParametersSetFactory;
 
 namespace DYN {
 
@@ -444,7 +442,7 @@ TEST(ModelerCommonTest, SetParameterFromPARFile) {
   const bool isInitParam = false;
 
   // Create a parameter set
-  boost::shared_ptr<parameters::ParametersSet> parametersSet = ParametersSetFactory::newInstance("Parameterset");
+  boost::shared_ptr<parameters::ParametersSet> parametersSet = boost::shared_ptr<ParametersSet>(new ParametersSet("Parameterset"));
   std::string string_value = "ok";
   parametersSet->createParameter("name_bool", false)
           ->createParameter("name_int", 2)
@@ -510,7 +508,7 @@ TEST(ModelerCommonTest, SetParametersFromPARFile) {
   const bool isInitParam = false;
 
   // Create parameter set
-  boost::shared_ptr<parameters::ParametersSet> parametersSet = ParametersSetFactory::newInstance("Parameterset");
+  boost::shared_ptr<parameters::ParametersSet> parametersSet = boost::shared_ptr<ParametersSet>(new ParametersSet("Parameterset"));
   const std::string string_value = "ok";
   parametersSet->createParameter("name_bool", false)
           ->createParameter("name_int", 2)
