+++ dynawo/sources/Modeler/DataInterface/test/TestStateVariable.cpp
@@ -58,7 +58,6 @@
 #include "DYNExecUtils.h"
 #include "DYNSubModelFactory.h"
 #include "DYNModelMulti.h"
-#include "PARParametersSetFactory.h"
 #include "DYNNetworkInterface.h"
 #include "DYNVoltageLevelInterface.h"
 #include "DYNShuntCompensatorInterface.h"
@@ -196,7 +195,7 @@ initializeModelNetwork(shared_ptr<DataInterface> data) {
   modelNetwork->initFromData(data);
   data->setModelNetwork(modelNetwork);
   modelNetwork->name("NETWORK");
-  shared_ptr<parameters::ParametersSet> parametersSet = parameters::ParametersSetFactory::newInstance("Parameterset");
+  shared_ptr<parameters::ParametersSet> parametersSet = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("Parameterset"));
   parametersSet->createParameter("bus_uMax", 0.);
   parametersSet->createParameter("capacitor_no_reclosing_delay", 0.);
   parametersSet->createParameter("load_alpha", 0.);
