+++ dynawo/sources/ModelicaCompiler/Scripts_OMC_1_13_2/headerPatternDefine.py
@@ -30,7 +30,6 @@ class HeaderPatternDefine:
 #include "DYNModelModelica.h"
 #include "DYNModelManagerCommon.h"
 #include "PARParametersSet.h"
-#include "PARParametersSetFactory.h"
 #include "PARParameter.h"
 #include "DYNSubModel.h"
 #include "DYNVariableForModel.h"
@@ -119,7 +118,6 @@ class HeaderPatternDefine:
 #include "DYNModelModelica.h"
 #include "DYNModelManagerCommon.h"
 #include "PARParametersSet.h"
-#include "PARParametersSetFactory.h"
 #include "PARParameter.h"
 #include "DYNSubModel.h"
 #include "DYNVariableForModel.h"
