+++ dynawo/sources/ModelicaCompiler/test/ModelicaModel/reference/GeneratorPQ_Dyn.cpp
@@ -612,7 +612,7 @@ boost::shared_ptr<parameters::ParametersSet> ModelGeneratorPQ_Dyn::setSharedPara
    // Propagating shared parameters default value 
 
    // This value may be updated later on through *.par/*.iidm data 
-  boost::shared_ptr<parameters::ParametersSet> parametersSet = parameters::ParametersSetFactory::newInstance("SharedModelicaParameters");
+  boost::shared_ptr<parameters::ParametersSet> parametersSet = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("SharedModelicaParameters"));
   int generator_NbSwitchOffSignals_internal;
   int generator_State0_internal;
