+++ dynawo/sources/ModelicaCompiler/test/ModelicaModel/reference/GeneratorPQ_Init.cpp
@@ -302,7 +302,7 @@ boost::shared_ptr<parameters::ParametersSet> ModelGeneratorPQ_Init::setSharedPar
    // Propagating shared parameters default value 
 
    // This value may be updated later on through *.par/*.iidm data 
-  boost::shared_ptr<parameters::ParametersSet> parametersSet = parameters::ParametersSetFactory::newInstance("SharedModelicaParameters");
+  boost::shared_ptr<parameters::ParametersSet> parametersSet = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("SharedModelicaParameters"));
 
   return parametersSet;
 }
