+++ dynawo/sources/ModelicaCompiler/test/ModelicaModel/reference_Test/Test_DynNoCalcVar.cpp
@@ -273,7 +273,7 @@ boost::shared_ptr<parameters::ParametersSet> ModelTest_Dyn::setSharedParametersD
    // Propagating shared parameters default value 
 
    // This value may be updated later on through *.par/*.iidm data 
-  boost::shared_ptr<parameters::ParametersSet> parametersSet = parameters::ParametersSetFactory::newInstance("SharedModelicaParameters");
+  boost::shared_ptr<parameters::ParametersSet> parametersSet = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("SharedModelicaParameters"));
   double a_internal;
   double b_internal;
