+++ dynawo/sources/Solvers/SolverIDA/test/TestBasics.cpp
@@ -56,7 +56,6 @@
 #include "DYNDynamicData.h"
 #include "PARParametersSet.h"
 #include "PARParameterFactory.h"
-#include "PARParametersSetFactory.h"
 #include "DYNTrace.h"
 #include "TLTimelineFactory.h"
 #include "DYNMacrosMessage.h"
@@ -67,7 +66,7 @@ boost::shared_ptr<Solver> initSolver() {
   // Solver
   boost::shared_ptr<Solver> solver = SolverFactory::createSolverFromLib("../dynawo_SolverIDA" + std::string(sharedLibraryExtension()));
 
-  boost::shared_ptr<parameters::ParametersSet> params = parameters::ParametersSetFactory::newInstance("MySolverParam");
+  boost::shared_ptr<parameters::ParametersSet> params = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("MySolverParam"));
   params->addParameter(parameters::ParameterFactory::newParameter("order", 2));
   params->addParameter(parameters::ParameterFactory::newParameter("initStep", 1.));
   params->addParameter(parameters::ParameterFactory::newParameter("minStep", 1.));
@@ -503,7 +502,7 @@ TEST(SimulationTest, testSolverIDAInit) {
   boost::shared_ptr<Model> model = initModelFromDyd("jobs/solverTestAlpha.dyd");
 
   // IDASVtolerances
-  boost::shared_ptr<parameters::ParametersSet> params = parameters::ParametersSetFactory::newInstance("MySolverParam");
+  boost::shared_ptr<parameters::ParametersSet> params = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("MySolverParam"));
   params->addParameter(parameters::ParameterFactory::newParameter("order", 2));
   params->addParameter(parameters::ParameterFactory::newParameter("initStep", 1.));
   params->addParameter(parameters::ParameterFactory::newParameter("minStep", 1.));
@@ -514,7 +513,7 @@ TEST(SimulationTest, testSolverIDAInit) {
   ASSERT_THROW_DYNAWO(solver->init(model, 0, 0), Error::SUNDIALS_ERROR, KeyError_t::SolverFuncErrorIDA);
 
   // IDASetMinStep
-  boost::shared_ptr<parameters::ParametersSet> params2 = parameters::ParametersSetFactory::newInstance("MySolverParam");
+  boost::shared_ptr<parameters::ParametersSet> params2 = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("MySolverParam"));
   params2->addParameter(parameters::ParameterFactory::newParameter("order", 2));
   params2->addParameter(parameters::ParameterFactory::newParameter("initStep", 1.));
   params2->addParameter(parameters::ParameterFactory::newParameter("minStep", -1.));
@@ -525,7 +524,7 @@ TEST(SimulationTest, testSolverIDAInit) {
   ASSERT_THROW_DYNAWO(solver->init(model, 0, 0), Error::SUNDIALS_ERROR, KeyError_t::SolverFuncErrorIDA);
 
   // IDASetMaxStep
-  boost::shared_ptr<parameters::ParametersSet> params3 = parameters::ParametersSetFactory::newInstance("MySolverParam");
+  boost::shared_ptr<parameters::ParametersSet> params3 = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("MySolverParam"));
   params3->addParameter(parameters::ParameterFactory::newParameter("order", 2));
   params3->addParameter(parameters::ParameterFactory::newParameter("initStep", 1.));
   params3->addParameter(parameters::ParameterFactory::newParameter("minStep", 1.));
@@ -536,7 +535,7 @@ TEST(SimulationTest, testSolverIDAInit) {
   ASSERT_THROW_DYNAWO(solver->init(model, 0, 0), Error::SUNDIALS_ERROR, KeyError_t::SolverFuncErrorIDA);
 
   // IDASetMaxOrder
-  boost::shared_ptr<parameters::ParametersSet> params4 = parameters::ParametersSetFactory::newInstance("MySolverParam");
+  boost::shared_ptr<parameters::ParametersSet> params4 = boost::shared_ptr<parameters::ParametersSet>(new parameters::ParametersSet("MySolverParam"));
   params4->addParameter(parameters::ParameterFactory::newParameter("order", -2));
   params4->addParameter(parameters::ParameterFactory::newParameter("initStep", 1.));
   params4->addParameter(parameters::ParameterFactory::newParameter("minStep", 1.));
